pyproject.toml
[project]
name = "fraud-detection"
version = "0.1.0"
description = "Fraud Detection System using ML and rule-based scoring"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask",
    "scikit-learn",
    "pandas",
    "numpy",
    "xgboost",
    "streamlit",
    "pyzbar",
    "Pillow",
    "python-dotenv",
    "requests",
    "tldextract",
]
